+++ aria/operations-for-logs/8.x/inspec/cassandra/controls/VLIC-8X-000014.rb
@@ -1,5 +1,5 @@
 control 'VLIC-8X-000014' do
-  title 'The Aria Operations for Logs Cassandra database must verify there are no user altered roles.'
+  title 'The VMware Aria Operations for Logs Cassandra database must verify there are no user altered roles.'
   desc  'In order to prevent unauthorized access organizations must ensure database roles are in their shipped state and have not been altered.'
   desc  'rationale', ''
   desc  'check', "
@@ -18,21 +18,21 @@ control 'VLIC-8X-000014' do
     If the output does not match the expected result, this is a finding.
   "
   desc 'fix', "
-    At the command prompt, run the following command for each unexpected \"member_of\":
+    At the command prompt, run the following command:
 
-    # /usr/lib/loginsight/application/lib/apache-cassandra-<VERSION>/bin/cqlsh-no-pass -e \"REVOKE <member of> FROM <role>;\"
+    # /usr/lib/loginsight/application/lib/apache-cassandra-<VERSION>/bin/cqlsh-no-pass -e \"DROP ROLE <ROLE>;\"
 
-    Note: Replace <member of> and <role> with the unexpected \"member_of\" and \"role\" values returned from the check.
+    Note: Replace <ROLE> with each unexpected role returned from the check.
   "
   impact 0.5
   tag severity: 'medium'
   tag gtitle: 'SRG-APP-000516-DB-000363'
-  tag satisfies: ['SRG-APP-000133-DB-000362']
+  tag satisfies: %w(SRG-APP-000133-DB-000362 SRG-APP-000378-DB-000365 SRG-APP-000516-DB-000363)
   tag gid: 'V-VLIC-8X-000014'
   tag rid: 'SV-VLIC-8X-000014'
   tag stig_id: 'VLIC-8X-000014'
-  tag cci: %w(CCI-000366 CCI-001499)
-  tag nist: ['CM-5 (6)', 'CM-6 b']
+  tag cci: %w(CCI-000366 CCI-001499 CCI-001812)
+  tag nist: ['CM-11 (2)', 'CM-5 (6)', 'CM-6 b']
 
   describe command("#{input('cassandraroot')}/bin/cqlsh-no-pass -e \"SELECT role, can_login, member_of FROM system_auth.roles;\"") do
     its('stdout.strip') { should match /lisuper\s*[|]\s*True\s*[|]\s*null/ }